requirements.txt
azure-eventhub==5.14.0
faker==23.3.0
pandas==2.0.3